requirements.txt
numpy==1.6.2
scipy==0.11
matplotlib==1.1.0
pandas=0.8.1
cloud_sptheme==1.4
wheezy.template
